Move DLE inspection/manipulation -> ir/iet/utils
