+++ devito/dle/__init__.py
@@ -1,4 +1,2 @@
-from devito.dle.inspection import *  # noqa
-from devito.dle.manipulation import *  # noqa
 from devito.dle.blocking_utils import *  # noqa
 from devito.dle.transformer import *  # noqa
