+++ devito/dle/backends/basic.py
@@ -8,11 +8,11 @@ import numpy as np
 from sympy import Symbol
 
 from devito.cgen_utils import ccode
-from devito.dle import retrieve_iteration_tree, filter_iterations
 from devito.dle.backends import AbstractRewriter, dle_pass, complang_ALL
 from devito.ir.iet import (Denormals, Expression, Call, Callable, List,
                            UnboundedIndex, FindNodes, FindSymbols,
-                           NestedTransformer, Transformer)
+                           NestedTransformer, Transformer,
+                           retrieve_iteration_tree, filter_iterations)
 from devito.symbolics import as_symbol
 from devito.tools import filter_sorted, flatten
 from devito.types import Scalar
