+++ devito/ir/iet/__init__.py
@@ -1,5 +1,5 @@
-from devito.ir.iet.utils import *  # noqa
 from devito.ir.iet.properties import *  # noqa
 from devito.ir.iet.nodes import *  # noqa
 from devito.ir.iet.visitors import *  # noqa
+from devito.ir.iet.utils import *  # noqa
 from devito.ir.iet.analysis import *  # noqa
