+++ devito/operator.py
@@ -11,7 +11,7 @@ from devito.arguments import infer_dimension_values_tuple
 from devito.cgen_utils import Allocator
 from devito.compiler import jit_compile, load
 from devito.dimension import Dimension
-from devito.dle import compose_nodes, filter_iterations, transform
+from devito.dle import transform
 from devito.dse import rewrite
 from devito.exceptions import InvalidArgument, InvalidOperator
 from devito.function import Forward, Backward, CompositeFunction
@@ -20,7 +20,7 @@ from devito.ir.clusters import clusterize
 from devito.ir.iet import (Element, Expression, Callable, Iteration, List,
                            LocalExpression, FindScopes, ResolveTimeStepping,
                            SubstituteExpression, Transformer, NestedTransformer,
-                           analyze_iterations)
+                           analyze_iterations, compose_nodes, filter_iterations)
 from devito.ir.support import Stencil
 from devito.parameters import configuration
 from devito.profiling import create_profile
