+++ tests/test_dle.py
@@ -9,12 +9,13 @@ from sympy import solve
 
 from conftest import EVAL
 
-from devito.dle import retrieve_iteration_tree, transform
+from devito.dle import transform
 from devito.dle.backends import DevitoRewriter as Rewriter
 from devito import Grid, Function, TimeFunction, Eq, Operator
 from devito.ir.iet import (ELEMENTAL, Expression, Callable, Iteration, List, tagger,
                            ResolveTimeStepping, SubstituteExpression,
-                           Transformer, FindNodes, analyze_iterations)
+                           Transformer, FindNodes, analyze_iterations,
+                           retrieve_iteration_tree)
 
 
 @pytest.fixture(scope="module")
